src-agents/phase1/requirements.txt
openai==1.43.0
pydantic==2.8.2
python-dotenv==1.0.1
tiktoken==0.7.0
azure-search-documents==11.5.1
azure-identity==1.17.1
uvicorn==0.30.6
fastapi==0.112.2
word2number==1.1